refactor(buildingStore): split into blockSlice, historySlice, particleSlice, achievementSlice, settingsSlice (#148)
